made update calendar button into separate component
